fix: Hash() now correctly uses unsigned long rather than unsigned int, and correctly xors bytes rather than bits

+++ include/parsing.h
@@ -8,7 +8,7 @@ typedef struct {
 
 int GetSlice(bufState* buf_state, Slice* slice, const char* delim, const int delim_len);
 
-unsigned int Hash(void* input, size_t size);
+unsigned long Hash(void* input, size_t size);
